dom/html/test/test_bug196523.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=196523
-->
<head>
  <title>Test for Bug 196523</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=196523">Mozilla Bug 196523</a>
<script>
  var expectedMessages = 2;
  SimpleTest.waitForExplicitFinish();
  window.addEventListener("message", function(e) {
    --expectedMessages;
    var str = e.data;
    var idx = str.indexOf(';');
    var val = str.substring(0, idx);
    var msg = str.substring(idx+1);
    ok(val == "true", msg);
    if (!expectedMessages) { SimpleTest.finish(); }
  }, false);
</script>
<p id="display">
  <iframe src="http://test1.example.org/tests/dom/html/test/bug196523-subframe.html"></iframe>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 196523 **/

</script>
</pre>
</body>
</html>
